Extract rustdoc-related code to `rustdoc` crate
