+++ crates/html_to_markdown/src/markdown.rs
@@ -23,7 +23,7 @@ impl HandleTag for ParagraphHandler {
             }
         }
 
-        match tag.tag.as_str() {
+        match tag.tag() {
             "p" => writer.push_blank_line(),
             _ => {}
         }
@@ -47,7 +47,7 @@ impl HandleTag for HeadingHandler {
         tag: &HtmlElement,
         writer: &mut MarkdownWriter,
     ) -> StartTagOutcome {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "h1" => writer.push_str("\n\n# "),
             "h2" => writer.push_str("\n\n## "),
             "h3" => writer.push_str("\n\n### "),
@@ -61,7 +61,7 @@ impl HandleTag for HeadingHandler {
     }
 
     fn handle_tag_end(&mut self, tag: &HtmlElement, writer: &mut MarkdownWriter) {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "h1" | "h2" | "h3" | "h4" | "h5" | "h6" => writer.push_blank_line(),
             _ => {}
         }
@@ -83,7 +83,7 @@ impl HandleTag for ListHandler {
         tag: &HtmlElement,
         writer: &mut MarkdownWriter,
     ) -> StartTagOutcome {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "ul" | "ol" => writer.push_newline(),
             "li" => writer.push_str("- "),
             _ => {}
@@ -93,7 +93,7 @@ impl HandleTag for ListHandler {
     }
 
     fn handle_tag_end(&mut self, tag: &HtmlElement, writer: &mut MarkdownWriter) {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "ul" | "ol" => writer.push_newline(),
             "li" => writer.push_newline(),
             _ => {}
@@ -131,7 +131,7 @@ impl HandleTag for TableHandler {
         tag: &HtmlElement,
         writer: &mut MarkdownWriter,
     ) -> StartTagOutcome {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "thead" => writer.push_blank_line(),
             "tr" => writer.push_newline(),
             "th" => {
@@ -158,7 +158,7 @@ impl HandleTag for TableHandler {
     }
 
     fn handle_tag_end(&mut self, tag: &HtmlElement, writer: &mut MarkdownWriter) {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "thead" => {
                 writer.push_newline();
                 for ix in 0..self.current_table_columns {
@@ -197,7 +197,7 @@ impl HandleTag for StyledTextHandler {
         tag: &HtmlElement,
         writer: &mut MarkdownWriter,
     ) -> StartTagOutcome {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "strong" => writer.push_str("**"),
             "em" => writer.push_str("_"),
             _ => {}
@@ -207,7 +207,7 @@ impl HandleTag for StyledTextHandler {
     }
 
     fn handle_tag_end(&mut self, tag: &HtmlElement, writer: &mut MarkdownWriter) {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "strong" => writer.push_str("**"),
             "em" => writer.push_str("_"),
             _ => {}
@@ -230,7 +230,7 @@ impl HandleTag for CodeHandler {
         tag: &HtmlElement,
         writer: &mut MarkdownWriter,
     ) -> StartTagOutcome {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "code" => {
                 if !writer.is_inside("pre") {
                     writer.push_str("`");
@@ -244,7 +244,7 @@ impl HandleTag for CodeHandler {
     }
 
     fn handle_tag_end(&mut self, tag: &HtmlElement, writer: &mut MarkdownWriter) {
-        match tag.tag.as_str() {
+        match tag.tag() {
             "code" => {
                 if !writer.is_inside("pre") {
                     writer.push_str("`");
