+++ crates/html_to_markdown/src/structure.rs
@@ -1,2 +1,1 @@
-pub mod rustdoc;
 pub mod wikipedia;
